Add JacksonRegistry
The new `JacksonRegistry` is an alternative to the `ProtoTypeRegistry`. It allows applications
to pass Java/Jackson objects as arguments to a CEL expression. This change does not support
_returning_ Java/Jackson objects _from_ a CEL expression.

The primary use case of this change is to enable CEL scripts to evaluate an expression against
some Java/Jackson object tree and return a CEL built-in type (primitive types like string or
boolean) - TL;DR security policy checks or filter-predicates.