Add strings extension functions with multiple overloads (#126)
#103 for multiple overloads:

- `indexOf`, providing new format with optional argument for starting position
- `lastIndexOf`, similar change as `indexOf`
- `replace`, providing optional argument for replace count
- `split`, providing optional argument for split count
- `substring`, providing optional argument for trailing range 
The javadocs and test cases were updated accordingly.

On implementation, `ProgramOptions` are updated to include these overloads, benefitting from https://github.com/projectnessie/cel-java/issues/110. 
The `replace` function is tricky as it may take 3 or 4 arguments which are represented as `FunctionOp`. The `FunctionOp` is made to execute different guard functions based on argument count.

The `replaceN` and `splitN` functionalities are new in Java. So porting them from Go.

Fixes #103