Fix Java 8 compatibiltiy
Although the CEL source code was always compatible with Java 8, the
compiled classes were not, because they were never build with
`--release 8`.

This change adds the changes to Java compilation.

CI didn't catch this issue bug, because the while build was run
with Java 8 and that caused the code to be _built_ again against
Java 8.

So this change also updates the CI workflow to run the tests against
the classes built by Java 11.