Enhance JacksonObjectT to return it's wrapped value
This is a small enhancement to JacksonObjectT to allow
it to return it's wrapped value (or itself).  This is
useful for having a CEL expr return the same object passed
in to the JacksonRegistry, such as when using a CEL
expr which may be an identity function under some
conditions.