Support Jackson 3 in addition to Jackson 2
The code is very much a copy of the code for Jackson 2, adapted in a few places to Jackson 3:

* Adapt to Jackson 3 type renames, especially for the `SerializationProvider` -> `SerializationContext` rename
* Adapt to the Jackson 3 mapper builder pattern and getting the required `SerializationContext` to inspect types in `Jackson3Registry`.

The public API type `org.projectnessie.cel.types.jackson3.Jackson3Registry` is in a different package and uses a different type name, although it results in a repetition of the Jackson major version in the type name.

The dependency-free `cel-standalone` artifact includes both Jackson 3 and Jackson 2 now.